Still working on adding validation to Edit Manufacturer and popup when you try to delete a manufacturer with items.

+++ src/app/services/manufacturer.service.ts
@@ -14,6 +14,7 @@ const httpOptions = {
 })
 
 export class ManufacturerService {
+  manufacturer: Manufacturer;
 
   constructor(private http:HttpClient) { }
 
@@ -31,7 +32,15 @@ export class ManufacturerService {
   }
 
   deleteManufacturer(id: number) {
-    return this.http.delete(`/server/manufacturers/${id}`);
+    this.getManufacturerById(id).subscribe(
+      data => {this.manufacturer = data},
+          err => console.error(err),
+          () => console.log('manufacturers loaded'));
+    if (this.manufacturer.items.length){
+      
+    } else {
+      return this.http.delete(`/server/manufacturers/${id}`);
+    }
   }
 
   updateManufacturer(id: number, manufacturer: Manufacturer) {
